ml-service/requirements.txt
flask>=2.3,<3.1
tensorflow==2.15.1
numpy<2.0
pillow>=10,<11
scikit-learn>=1.3,<1.6
protobuf>=3.20.3,<5.0.0
